Prefer the package that matches the module name for indexing.
A given release reports a list of packages shipped in its modules.

However, due to cuckoo package, sometimes a given package
can report to be both authorized and indexed more than once.

And in the case of the cuckoo package, it can report a different, or
even undefined version to be "indexed" and "authorized".

Subsequently, in such a case that we detect a given release indicating
more than one package name is indexed, if there was any module shipping
that package name with an identical module name, we prefer that ones
version.

Hence:

  Foo.pm -> Foo=1.0, Bar=undef
  Bar.pm -> Bar=1.0

Should be:

  Foo=1.0, Bar=1.0

Not:

  Foo=1.0, Bar=undef

See also metacpan/metacpan-api#511